# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: passwords
    description: Password generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /passwords/generate:
    post:
      operationId: generatePasswords
      tags: [passwords]
      summary: Generate smart password suggestions
      description: Analyzes previous passwords and returns memorable, strong suggestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePasswordsBody"
      responses:
        "200":
          description: Password suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratePasswordsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GeneratePasswordsBody:
      type: object
      properties:
        previousPasswords:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 3
          description: List of up to 3 previous passwords
      required:
        - previousPasswords
    PasswordSuggestion:
      type: object
      properties:
        password:
          type: string
        strength:
          type: string
          enum: [weak, fair, strong, very_strong]
        memorabilityNote:
          type: string
          description: Human-readable note on what makes it memorable
        strengthNote:
          type: string
          description: Human-readable note on what makes it strong
      required:
        - password
        - strength
        - memorabilityNote
        - strengthNote
    GeneratePasswordsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/PasswordSuggestion"
        analysisNote:
          type: string
          description: Summary of patterns found in the previous passwords
      required:
        - suggestions
        - analysisNote
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error