# lib/api-zod/src/generated/types/passwordSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordSuggestionStrength } from "./passwordSuggestionStrength";

export interface PasswordSuggestion {
  password: string;
  strength: PasswordSuggestionStrength;
  /** Human-readable note on what makes it memorable */
  memorabilityNote: string;
  /** Human-readable note on what makes it strong */
  strengthNote: string;
}